Remove other dead code
